bot/requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
sqlalchemy==2.0.36
redis==5.2.0
python-dotenv==1.0.1
hypothesis==6.119.0
pytest==8.3.4
pytest-asyncio==0.23.8
aiohttp<3.11